Materials now "Burn", if they are burning, and go through the stages not burnt, slightly burnt, moderately burnt and heavily burnt.
This approach uses a mixture of Spoiling which needs a Be() for "burning" things over time and Rusting, which stores the rust data for a material efficiently. There was an earlier bug because of missing line in bodypart.h:
virtual truth AddBurnLevelDescription(festring&, truth) const { return false; }

This paves the way for the bitmaps to be manipulated as per RustData.
There is also a change to the way items become burnt. The material gets burnt instead. This means it is now conceptually correct.

Known issue:
With the changes to the Burn level description, things are no longer piling properly anymore, so this needs to be attended to at some point.