Fixed window bug
I couldn't find the root cause of the bug, but I was able to add a check
to see if the tile had a window or not. This may cause problems if there
are walls that need to be able to see under, but I don't think that is a
likely case. Fixes #3