Fixed a bug that could crash the game when reading a scroll of repair
...while carrying a corpse. This is probably not a good solution since
shouldn't corpses be able to burn? Note that IsRusted is overridden to
return false for corpses too, so it seems corpses can't rust either.

Cause of crash: Without this corpse::IsBurnt override, the game would
call item::IsBurnt which accesses the MainMaterial, however corpses
don't seem to have a MainMaterial (the MainMaterial pointer was 0 when
I debugged this).