Incorporating the fire system with the god system, round 1:
Dulcis will now put out flames on you and your companions if your body parts are on fire. Your pets must be in adjacent squares to you.
This is prioritised over taming, so look out, it is big change!

Silva will check if you or any items in your inventory/equipment are burning. If so, then she will provide a rain shower over your head.
Spread your burning items (books and scrolls) out over this area to allow the rain to fight the fires! Don't stack items, or there is less rain water to go around.
Players with burning bodyparts will also benefit from this, as the rain washes over bodyparts.
Silva also prioritises this over earthquakes and wolves, so this is another big change.

Player gets an extinguish message when something extinguishes and it is the last visible light source.

A minuscule change, but FightFire now decrements thermal energy according to the square root of the fluid volume, rather than volume/10. This works far, far better now especially when there are only small amounts of liquid.
As a result, things get put out a lot faster for small amounts of fluid, so less stressing-out for the player.
Beware, large explosions will still cause fires requiring higher amounts of fluid with which to fight fires, owing to higher accumulated energy.