Fire and god systems remix, part two:
Seges will now check to see whether the player has burnt limbs and will restore these to an unburnt state. This is lower down in the priority list, just after satiation.

Loricatus has two effects:
First priority is now to try to give you a light source in the dark. If player prays to Loricatus in the dark, he will try to set on fire whatever you are wielding. Effect is Relation / 2, so you can potentially burn down dragon hide if you're not careful. This represents a somewhat big change to this prayer effect: if it's not an emergency and you don't need the extra light source, don't pray to Loricatus in the dark.
Will remove burns on equipped burnt gear, without removing flames. Put fires out first for best effect. This is placed at last priority, behind fixing broken gears.

Made a way of checking whether a character is on fire.
Discovered that RemoveBurns() is a little antiquated in its present state.
Discovered that the fire messaging system needs a tough look.

Fixed: scrolls of repair which unintentionally put flames out.
Scroll of repair is updated to repair burnt artificial limbs.

A little info about these functions in their present state:
void RemoveBurns(); // removes the burns and restores HP, without resetting the burn level counter, which is kind of silly
void ResetThermalEnergies(); // resets a material's stored energy values, which causes a material to extiguish - it calls Extinguish(false) (false => no messages)
ResetBurning(); // Does roughly the same as RemoveBurns(), only without the message(?) and it resets the burn level counter, and no HP restore