Make 'open', 'close' and 'use key' more context-aware
For example, if the player is standing next to a door, then the ‘open’,
‘close’ and ‘use key’ commands will automatically act on that door
without asking for a direction.

However if there’s also, say, a chest next to the player (in addition
to the door), then the commands will ask whether the player wants to
act on the door or the chest, if (and only if) it’s ambiguous which one
the player means. For example if the door is already open, then the
‘open’ command will open the chest and ignore the door completely.