Fix out-of-bounds access of MoveVector with RANDOM_DIR
game::GetMoveVector was occasionally called with RANDOM_DIR (9),
which would cause an out-of-bounds read.

This wasn't causing anything noticeable, probably because it would
just access the first element of the next static array instead. Although
in that case the result was not random, which was the whole point
of RANDOM_DIR.

(A simple assert() would have prevented this...)