Fix buffer overflow when calling memcmp on C-strings
If Result and CStr were identical and N was sufficiently large (> index
of '\0' + 1) then memcmp would access past the null terminator. This
was probably never the case but at least now this won't trigger an
error with AddressSanitizer. With strncmp it will do the same except
stop when it encounters a null terminator.