Put largest member first in blitdata union
This allows the union to be initialized without narrowing the
initializing expression w.r.t. size, which could result in a bug on
platforms with 16-bit int. (Because union initializations in C++ can
only initialize the first member.) Note that an int initializer must
still be sign-converted to col24, but this shouldn't affect anything.

This also makes the initializing expressions more readable by average,
since most of the time the union is initialized with col24, and not
int. It is initialized with an int only once.