Lanterns ignite square they break in
There was a seg fault throwing a lantern at genetrix so I threw in lots
of debug printing to make sure this functions. Also, 100 fire damage is
almost certainly too much. Move it to a variable on Lantern later.

Perhaps redo this to use an explosion and change explosions to have
option to do only FIRE damage?