Add toggle option for the smart open/close/apply
as discussed in Attnam/ivan#85. This can be used to disable the
"smart" behavior introduced in Attnam/ivan#41, and instead use the
old, simpler, behavior.

Sorry for the horrible diff. All I did was indent the code for the
"smart" behavior (to put it in an 'if'), and add back the old code
that was removed in commit ef55864.