IVAN automatically causes scripted locations to appear in the world map
Two sets of data: a list of available locations (AvailableLocations); and, a list of places (ToBePlaced) that need to appear on the world map.

So far, IVAN puts places on the PetrusLikes continent only. There are various limitations. Checking for underflow and overflow between ToBePlaced and AvailableLocations in the placement routine is unimplemented. Places can be overwritten by the core-location placement method (Gloomy Caves etc).
The vector containing the AvailableLocations is sorted in order of manhattan distance to Attnam's position vector.
At present, it looks like the locations should fill up in the PetrusLikes continent; once it is full, the routine should move to another continent, or fill up positions according to nearest, irrespective of continent. So many options, but which to choose? And which would be most extensible?

There is some code for investigating the statistics of the ground world terrain types that come up due to the poisson disc algorithm.