Code tidy-up
Removed sandbox terrain analysis.
Restored PetruLikes to be between 25 - 1000 tiles big, as in 0.50 and CVS.
Made the language in the worldmap comments more consistent between terms like position, location and place: Position is a co-ordinate, place is a user-defined dungeon, location is a position on the worldmap where we can put a place.