Changes are to script only, to create a 'bleeding edge' version to tweak the Tomb of Xinroch.
Prevent resurrected Xinroch from being generated outside the boss level.
Removed altar generation from mazes in dungeons/XinrochTomb.dat.
Removed the neutral necromancer in the necro-chamber because his summons were getting in the way.
Locked doors and made some layout changes to the dwarven gas chamber. This is to make it easier for IVAN to place the room on the level.
Fixed ground terrains under wall squares to be the same type as floor squares in maze rooms.
Changed ebony floors with octiron walls to bone floors with octiron walls, to improve contrast.
Added some light sources to the tactical level to make better use of the enhanced line of sight.
Fountains in the ice cave are now frozen.
Minimum room size is now 4x4 instead of 3x3.
Music starts on level 2 now instead of on the tomb entrance level.

Changes requiring a new build will be ready by December release.