Better kobolds
Rather than chieftains and lords, I renamed higher-level kobolds to
hunters and warriors. Kobolds seem to be a tribal society and it sounds
weird to send a large numbers of chieftains to battle, not even talking
about how "lord" is pretty unappropriate for a tribal leader.

Kobold assassins added for some extra fun. :)