Add guardian to UT vault
Right now, the vault is free items for anyone who figures out there is
something hidden there. IVAN is not a game for free items.

Oak golem should be enough to guard the treasure and yet be defeatable
by the player early enough for the vault to be useful even the first
time player passes through UT.