More interesting final TX level (#390)
Resurrected Xinroch's power is fine, but IMHO his level is a bit
lacking. Compare with Oree's lair, which has a design unseen anywhere
else in the game.

I suggest using a reversed TX Entry temple, ie. a reversed pentagram
where Xinroch can stand and fight. Fake UT1 entry for a bit of fun with
the player, then a fight in open area around the temple, and finally a
showdown with Xinroch in ruby-and-diamond temple.

With the new materials at our disposal, I changed the earth and wall
materials for Xinroch ghost and resurrected Xinroch levels a bit to have
a nice color scheme. :)