Add resistance shields
There are already other options for gaining resistances, but I think
more options are welcome. They allow you to build different kits, not
requiring that a certain slot whould always be used for a resistance
item. Sometimes you want a ring, sometimes you want a cloak, sometimes
you want a shield.

Of course, they need to be rare enough that you are not guaranteed to
have a resistance, but I hopefully set their frequency low enough.