Add artifact mask
In an age long past, the goblin empire of Quetzaltia fought a terrible
war against the spawn of Scabies, the dreaded mind worms. The goblins
were barely holding on until the wizards and alchemists of their king,
the venerable Tlachtlallotlachitl, invented a new type of mask that
rendered the powers of mind worms useless. Very few of those masks
remain today, but any worm-slayer would give a fortune to own one.