Add Tweraif Freedom victory
You can speak with Kaethos and Terra after you liberate New Attnam for a
bit of a backstory. Then kill Lobh-se and drop the seedling on the
banana landing spot.

There is also an alternate way to secure the freedom of Tweraif. If you
have 50 000 gp to spare, go to Thesaurius (the hoardmaster in the
Cathedral), and you can bribe him so that you will become the new
viceroy and make the life in New Attnam better.