Bugfixes (#518)
* Fix for PR #382

Fixes issue #511, replacing first person in crafting messages with
second person, as in the rest of the game.

* Fix for PR #492

Apparently, TweraifIsFree() only works in NA, so Terra was not saying
what she should.

* Crafting wording

Simplify wording of crafting options. If @AquariusPower disagrees, I can
revert this, but I think some of the options were worded in a bit
strange manner.