Create CONTRIBUTING.md
I've felt for some time now that we could have a guideline for contributing. Here's my suggestion for a contribution guideline, based on my observation of how how people have been contributing over the years. It's not exhaustive by any means, please feel free to make any changes based on your own observations and collaboration experience. As usual, there is no time frame for completion.

One feature I would like to see is if a PR is approved, then anyone including the author can merge that PR. Not sure if this is useful, but it feels like this builds consensus and is more laid back. I don't know how anyone else feels about merging ones own work, but maybe there are circumstances which merit this?

I think we can mention the useful feature of GitHub's draft PR to distinguish between work in progress and stuff that is ready to be reviewed/merged.

There's also a things to do, a job list, and a forbidden jobs. Upsetting the balance is probably the worst, after coding bugs. I think there are probably enough materials and weapons at this stage in development.