Hide black market
After some play testing, I realized that wandering on the worldmap 
looking for the black market is not really working out. Unfortunately it 
tends to spawn either close to one of the other dungeons, leading to it 
being discovered rather soon, or in very remote location, even on other 
continents.

Because the black market should remain a late game area, I think it 
would be better to make it initially hidden and reveal it only once the 
player can buy a map that would lead them there.

BTW, why was Tomb of Xinroch "locationAW"? :)