Black market (#562)
* Merge remote-tracking branch 'upstream/master'

* Fixes

Mirrored items cannot be dismantled into permanent lump of material #534

Use fix from @AquariusPower to stop weapon swap from auto-stealing items 
from shops #555

* Implement Black Market

Especially in later stages of the game, players can find themselves with 
large amounts of money that can no longer be spend on anything useful 
(barring priest services and fixing equipment in Attnam), as shops not 
always offer good items, or anything of value was already bought.

This adds a new dungeon, the Black Market, which is randomy placed on 
the worldmap and may be found by players who go explore the wilderness 
(This also gives some incentive to go explore, as otherwise the player 
will only go for the shortest route between UT, Attnam and GC/TX). It 
has two levels: The first one is full of guards, the second has a hige 
shop of overpriced items.

The purpose of this is to have a palce where players can spend all their 
hard-earned money, with enough items generated to ensure there will be 
at least something for every player. However, to prevent unbalancing the 
game with new shopping options, the Black Market has intentionally 
overpriced items that will only be affordable to players who have 
accumulated some excess funds.

* Tweaks and fixes

* Add missing NEWS

* Minor changes and fixes

* Hide black market

After some play testing, I realized that wandering on the worldmap 
looking for the black market is not really working out. Unfortunately it 
tends to spawn either close to one of the other dungeons, leading to it 
being discovered rather soon, or in very remote location, even on other 
continents.

Because the black market should remain a late game area, I think it 
would be better to make it initially hidden and reveal it only once the 
player can buy a map that would lead them there.

BTW, why was Tomb of Xinroch "locationAW"? :)

* Fix wand of striking in GC