Quest changes
Changes the structure of quests to better allow for multiple alternative 
main quests without having to juggle with the encrypted scroll for all 
of them.

Petrus now takes the encrypted scroll and gives the player some time to 
wander Attnam. If the player chats with him again, they will receive the 
GC quest, but they can also find other people to get other quests.

Chat with the necromancer for TX quest, plus some preliminary work for 
Aslona quests is done.