Worldmap experiments (#631)
* Allow location placement to re-roll in worldmap generation

Without this fix, attempting to use the continue statement introduced in this commit would cause some locations to appear in the ocean occasionally. The trick is not to spawn() until it is certain that there is enough room for all the locations. The clues were all there.

* Add the FastNoise library

Use FastNoise to seed the world map using Simplex noise

* Configurable world sizes, shapes, landmasses

* Add world seed configuration option

* Add 4D Simplex Noise to FastNoise library

* Put FastNoise in its own folder

* Limit LandType to Pangea and Continents

Also added missing 96x96 world size.

* Show the world seed via `^` command

* Altitude map from 4D simplex noise

Work in progress...

* Refactor periodic noise generator codes

Scaling appears to be inherent, irrespective of aspect ratio.

* Allow random seed by requesting WorldSeed=-1

* Add description of world shape

Divulged in in-game dialogue with the Priest of Valpurus.
Also added a tweak to make the the Underwater Tunnel Exit the centre of worldmap location placement.

* Player movement wrap through world map boundaries

Depending on the shape of the world. Cylindrical worlds player movement can wrap through x boundaries, toroidal worlds player movement can wrap through x and y boundaries of the world map. Only allows movement if terrain is walkable on the other side.

* Connect latent features into the generator

Select world seed, and generate with this seed.
Allow land type selection to affect noise frequency.
Player movement wrapping in the world map, according to world shape.

* Remove memory leaks from Poisson disc sampler

* Re-sample if core locations not on same continent

Count the number of forced world regenerations, make the sample radius dependent on this number.

* Implement ForcePlacementOnAnyTerrain

Kill a bunch of segfaults by making the small world size 49x49 instead of 48x48, which was triggering all kinds of insane bugs. Not ideal, but seems to work.

* Add in remaining checks for bad initial seed

* Reduce number of generic locations to 8

* Make wizmode worldmap reveal respect owterra.dat

* Fine tuning noise frequency w.r.t. world size

Final fine-tuned values for noise frequencies.
Also fixed a bug where the screen does not get re-drawn for the one-screen world size.
Add generation limit for bad seeds, including bad seed detection and message.

* Make NativeGTerrainTypes into a fearray<int>

Allowing multiple terrain types for each dungeon helps to expedite dungeon generation.
Take the additional edits to owterra.dat with a grain of salt, they will be rolled back.

* Patch up previous screen re-draw fix

* Refactor out spurious code

* Propose Pyramid appear in desert or jungle tile

* Change the default world size to small