Fix identical mob names
Controlled polymorph requires that you specify the name of your target 
form, but there are several mobs with identical names and the game 
doesn't know how to deal with that. Fix being unable to polymorph into 
sirens and dark knights by making their names distinct.

This has an incidental worldbuilding side-benefit, as we now know that 
sirens are divided into nine clans, maybe one for each of the nine Grand 
Trees? ;)