Minor bugfixes (#649)
* Merge remote-tracking branch 'upstream/master'

* Diets

* Werewolfwolves now have canine diet.
* Add magic materials to imp diet, because they already eat weird 
materials.
* Buff beer nutrition to promote alcoholism.

Fixes #630

* Fix identical mob names

Controlled polymorph requires that you specify the name of your target 
form, but there are several mobs with identical names and the game 
doesn't know how to deal with that. Fix being unable to polymorph into 
sirens and dark knights by making their names distinct.

This has an incidental worldbuilding side-benefit, as we now know that 
sirens are divided into nine clans, maybe one for each of the nine Grand 
Trees? ;)

* Fix archangels

* Gladius should not have a PERMANENT RUBY TWO-HANDED SCIMITAR
* Amatrix needed a buff
* Xinroch can actually be friendly

* New NEWS

* Switch inventory to selectable list

This allows for use of F1 help in the main inventory window, so the 
players can more easily read item descriptions.