feat: Make more new options default
We have quite a few new options, yet many are disabled by default. A lot of players will never give the Options menu much thought, so they will never discover some nifty quality of life additions that are in the game for many years now. This attempts to set all options that improve QoL but do not alter base IVAN experience as enabled by default.

Full screen is on by default now. I have seen many players complain about the small window and I think those who do NOT use fullscreen are in minority.

Various extra info and better handling options are enabled.

Graphics scaling is NOT enabled by default, because while nice, it radically alters the game's looks, in a way that I feel should not be the default. Old vanilla IVAN graphics still looks the best!

StartWithNoPet became StartWithPet. It is a bool option and there is no need to make it negative and thus confusing.

Also reorganizes the Options menu to group similar options and to better fit each section on each page.